requirements.txt
Flask
flask-cors
numpy
opencv-python
Pillow